# Optimise query for platform-admin/live-services and trial-services